Organized models, and added abstract account class
