+++ angular-client/src/app/models/account.class.ts
@@ -1,0 +1,16 @@
+import {Address} from './address';
+import {IAccount} from './interfaces/account';
+
+export abstract class Account implements IAccount {
+  accountId: string;
+  firstName: string;
+  lastName: string;
+  email: string;
+  password: string;
+  address: Address;
+  phone: string;
+
+  constructor() {
+
+  }
+}
