--- angular-client/src/app/models/account.ts
@@ -1,21 +1,0 @@
-import {Address} from './address';
-
-export interface IAccount {
-  accountId: string;
-  firstName: string;
-  lastName: string;
-  email: string;
-  password: string;
-  address: Address;
-  phone: string;
-}
-
-export class Account implements IAccount {
-  accountId: string;
-  firstName: string;
-  lastName: string;
-  email: string;
-  password: string;
-  address: Address;
-  phone: string;
-}
