+++ angular-client/src/app/models/address.ts
@@ -1,12 +1,4 @@
-export interface IAddress {
-  streetLine1: string;
-  streetLine2: string;
-  postalCode: string;
-  province: string;
-  city: string;
-}
-
-export class Address implements IAddress {
+export class Address {
   streetLine1: string;
   streetLine2: string;
   postalCode: string;
