+++ angular-client/src/app/models/appointment.ts
@@ -1,7 +1,8 @@
 import {Patient} from './patient';
 import {Clinic} from './clinic';
+import {AppointmentType} from './enums/appointment.enum';
 
-export interface IAppointment {
+export class Appointment {
   appointmentId: number;
   clinic: Clinic;
   reason: string;
@@ -11,23 +12,4 @@ export interface IAppointment {
   endTime: Date;
   type: AppointmentType;
   patient: Patient;
-}
-
-export class Appointment implements IAppointment {
-  appointmentId: number;
-  clinic: Clinic;
-  reason: string;
-  preferredDate: Date;
-  preferredTime: Date;
-  startTime: Date;
-  endTime: Date;
-  type: AppointmentType;
-  patient: Patient;
-}
-
-export enum AppointmentType {
-  CONFIRMED,
-  REQUESTED,
-  CANCELLED,
-  COMPLETED,
 }
