+++ angular-client/src/app/models/clinic.ts
@@ -1,17 +1,9 @@
-import {Appointment} from "./appointment";
+import {Appointment} from './appointment';
 import {Address} from './address';
 import {HealthPractitioner} from './healthPractitioner';
 import {MedicalAdmin} from './medicalAdmin';
 
-export interface IClinic {
-  address: Address;
-  name: string;
-  medicalAdmins: MedicalAdmin[];
-  healthPractitioners: HealthPractitioner[];
-  appointments: Appointment[];
-}
-
-export class Clinic implements IClinic {
+export class Clinic {
   address: Address;
   appointments: Appointment[];
   healthPractitioners: HealthPractitioner[];
