+++ angular-client/src/app/models/enums/appointment.enum.ts
@@ -1,0 +1,6 @@
+export enum AppointmentType {
+  CONFIRMED,
+  REQUESTED,
+  CANCELLED,
+  COMPLETED,
+}
