+++ angular-client/src/app/models/interfaces/account.ts
@@ -1,0 +1,12 @@
+import {Address} from '../address';
+
+export interface IAccount {
+  accountId: string;
+  firstName: string;
+  lastName: string;
+  email: string;
+  password: string;
+  address: Address;
+  phone: string;
+}
+
